Refactored crop generator
